+++ CHANGELOG.md
@@ -43,3 +43,8 @@
 * feat(BetterTerminal): add separate enable and smooth GUI option for each type of terminal
 * fix(BetterTerminal): fix being unable to queue Order
 * fix(WindowProperties, RawInput): fix mouse not centered when opening GUI
+
+## 0.8.1 (UTC+8 2025/04/07 23:30)
+
+* fix(WindowProperties, RawInput): fix mouse not centered when opening GUI
+* feat(YCLeapMenu): add support for Haunt